+++ app/presenters/hyrax/image_presenter.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 module Hyrax
   class ImagePresenter < ::Spot::BasePresenter
+    include ::Spot::ExportsImageDerivatives
+
     delegate :contributor, :creator, :date, :date_associated,
              :date_scope_note, :description, :donor, :inscription,
              :keyword, :language_label, :local_identifier, :note,
@@ -10,14 +12,6 @@ module Hyrax
              :rights_holder, :rights_statement, :standard_identifier,
              :subtitle, :title_alternative,
              to: :solr_document
-
-    def manifest_metadata_fields
-      %i[
-        title subtitle title_alternative creator contributor date
-        description inscription keyword language_label location
-        subject subject_ocm standard_identifier rights_statement
-      ]
-    end
 
     def subject_ocm
       solr_document.subject_ocm.sort