+++ app/services/spot/mappers/cpw_shashinkai_mapper.rb
@@ -50,9 +50,7 @@ module Spot::Mappers
     #
     # @return [Array<RDF::Literal>]
     def title
-      metadata['title.english']
-        .select { |v| identifier_prefix?(v) }
-        .map { |v| RDF::Literal(v, language: :en) }
+      titles_with_identifiers
     end
 
     # relies on LanguageTaggedTitles to gather up our usual title_alternative suspects
@@ -60,16 +58,7 @@ module Spot::Mappers
     #
     # @return [Array<RDF::Literal>]
     def title_alternative
-      super + metadata['title.english'].reject { |v| identifier_prefix?(v) }.map { |v| RDF::Literal(v, language: :en) }
+      super + titles_without_identifiers
     end
-
-    private
-
-      # Does a value look like "[aa0001] title" ?
-      #
-      # @return bool
-      def identifier_prefix?(value)
-        value.match?(/^\[\w{2}\d{4}\]/)
-      end
   end
 end