+++ config/initializers/spot_overrides.rb
@@ -10,7 +10,6 @@ Rails.application.config.to_prepare do
   Hyrax::Dashboard::CollectionsController.include Spot::CollectionsControllerBehavior
 
   Hyrax::CollectionsController.presenter_class = Spot::CollectionPresenter
-  Hyrax::CollectionsController.include Spot::CollectionSlugsBehavior
   Hyrax::CollectionsController.include Spot::CollectionsControllerBehavior
 
   Hyrax::DerivativeService.services = [
@@ -18,7 +17,8 @@ Rails.application.config.to_prepare do
     ::Hyrax::FileSetDerivativesService
   ]
 
-  # see above
+  Hyrax::CurationConcern.actor_factory.swap(Hyrax::Actors::CollectionsMembershipActor, Spot::Actors::CollectionsMembershipActor)
+
   Hyrax::ContactFormController.class_eval { layout 'hyrax/1_column' }
 
   # the dashboard/my/collections (+ thus, dashboard/collections) controller defines