+++ config/routes.rb
@@ -19,7 +19,6 @@ Rails.application.routes.draw do
   mount BlacklightAdvancedSearch::Engine => '/'
   mount Hydra::RoleManagement::Engine => '/admin'
   mount Hyrax::Engine, at: '/'
-  mount Riiif::Engine => 'images', as: :riiif if Hyrax.config.iiif_image_server?
   mount Sidekiq::Web => '/sidekiq'
   mount Qa::Engine => '/authorities'